requirements.txt
streamlit==1.28.1
openai==1.101.0
python-dotenv==1.0.0
